Add yaml param to publish.

+++ entrypoint.sh
@@ -67,33 +67,35 @@ fi
 JEKYLL_ENV=production bundle exec ${BUNDLE_ARGS} jekyll build -s ${GITHUB_WORKSPACE}/${JEKYLL_SRC} -d build ${VERBOSE}
 echo "Jekyll build done"
 
-cd build
+if [ "${JEKYLL_PUBLISH}" = true ]; then
+  cd build
 
-# No need to have GitHub Pages to run Jekyll
-touch .nojekyll
+  # No need to have GitHub Pages to run Jekyll
+  touch .nojekyll
 
-# Is this a regular repo or an org.github.io type of repo
-case "${GITHUB_REPOSITORY}" in
-  *.github.io) remote_branch="master" ;;
-  *)           remote_branch="gh-pages" ;;
-esac
+  # Is this a regular repo or an org.github.io type of repo
+  case "${GITHUB_REPOSITORY}" in
+    *.github.io) remote_branch="master" ;;
+    *)           remote_branch="gh-pages" ;;
+  esac
 
-if [ "${GITHUB_REF}" = "refs/heads/${remote_branch}" ]; then
-  echo "::error::Cannot publish on branch ${remote_branch}"
-  exit 1
+  if [ "${GITHUB_REF}" = "refs/heads/${remote_branch}" ]; then
+    echo "::error::Cannot publish on branch ${remote_branch}"
+    exit 1
+  fi`
+
+  echo "Publishing to ${GITHUB_REPOSITORY} on branch ${remote_branch}"
+  echo "::debug::Pushing to https://${JEKYLL_PAT}@github.com/${GITHUB_REPOSITORY}.git"
+
+  remote_repo="https://${JEKYLL_PAT}@github.com/${GITHUB_REPOSITORY}.git" && \
+  git init && \
+  git config user.name "${GITHUB_ACTOR}" && \
+  git config user.email "${GITHUB_ACTOR}@users.noreply.github.com" && \
+  git add . && \
+  git commit -m "jekyll build from Action ${GITHUB_SHA}" && \
+  git push --force $remote_repo master:$remote_branch && \
+  rm -fr .git && \
+  cd .. 
 fi
-
-echo "Publishing to ${GITHUB_REPOSITORY} on branch ${remote_branch}"
-echo "::debug::Pushing to https://${JEKYLL_PAT}@github.com/${GITHUB_REPOSITORY}.git"
-
-remote_repo="https://${JEKYLL_PAT}@github.com/${GITHUB_REPOSITORY}.git" && \
-git init && \
-git config user.name "${GITHUB_ACTOR}" && \
-git config user.email "${GITHUB_ACTOR}@users.noreply.github.com" && \
-git add . && \
-git commit -m "jekyll build from Action ${GITHUB_SHA}" && \
-git push --force $remote_repo master:$remote_branch && \
-rm -fr .git && \
-cd .. 
 
 exit $?
